Added program "Outliner"
Wrote a small program that lets you calculate the sum, average and median of data sets. So far I have only been testing this on very small amounts of data but it does work as intended, but it may or may not work with large data sets, that remains to be seen...

I also made it print out tags alongside each output, this way you can have a tag called "Weight" and underneath it you would have the sum of all weights, then the average of all the weights and then you would have the median of them all as well.